% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{build_block_df}
\alias{build_block_df}
\title{Build dataframe with block location data}
\usage{
build_block_df(json)
}
\arguments{
\item{json}{a json file from DAI}
}
\value{
a dataframe
}
\description{
Build dataframe with block location data
}
\examples{
\dontrun{
build_block_df("pdf_output.json")
}
}
