% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{build_token_df}
\alias{build_token_df}
\title{Build dataframe with token location data}
\usage{
build_token_df(json)
}
\arguments{
\item{json}{a json file from DAI}
}
\value{
a dataframe
}
\description{
This function takes a json output file from Google Document AI (DAI)
and builds a dataframe with page location data for all identified tokens.
}
\details{
the dataframe has tokens as rows and location variables as columns
(start index, end index, left boundary, right boundary, top boundary, and bottom boundary.)
}
\examples{
\dontrun{
build_token_df("pdf_output.json")
}
}
