% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_folder}
\alias{create_folder}
\title{Create folder in Google Storage}
\usage{
create_folder(fname, bucket = Sys.getenv("GCS_DEFAULT_BUCKET"))
}
\arguments{
\item{fname}{name of new folder}

\item{bucket}{name of bucket. Defaults to global bucket if set.}
}
\description{
Create folder in Google Storage
}
\details{
Creates a "folder" in your Google Storage bucket by uploading an empty file.
}
\examples{
\dontrun{
create_folder("results")
create_folder("pdfs/documents/json_output")
}
}
