% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dai_auth}
\alias{dai_auth}
\title{Get service account token}
\usage{
dai_auth(
  scopes = "https://www.googleapis.com/auth/cloud-platform",
  path = Sys.getenv("GCS_AUTH_FILE")
)
}
\arguments{
\item{scopes}{GCS auth scopes for the token.}

\item{path}{path to a json file with a service account key.}
}
\description{
Get service account token
}
\examples{
\dontrun{
dai_auth()
}
}
