% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_to_dai.R
\name{dai_status}
\alias{dai_status}
\title{Check the status of an asynchronous processing job}
\usage{
dai_status(response, loc = "eu", token = dai_token(), verbose = FALSE)
}
\arguments{
\item{response}{An HTTP response object generated by \code{dai_async()}.}

\item{loc}{A two-letter region code; "eu" or "us".}

\item{token}{An authentication token generated by \code{dai_auth()} or another auth function.}

\item{verbose}{Whether to output the full response; boolean.}
}
\value{
An HTTP response (if verbose = TRUE).
}
\description{
Check the status of an asynchronous processing job
}
\examples{
\dontrun{
# Short status message:
response <- dai_async(myfiles)
dai_status(response)

# Full status details:
response <- dai_async(myfiles)
status <- dai_status(response, verbose = TRUE)
}
}
