% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_to_dai.R
\name{dai_sync_tab}
\alias{dai_sync_tab}
\title{OCR page synchronously and extract table data}
\usage{
dai_sync_tab(file, proj_id = get_project_id(), loc = "eu", token = dai_token())
}
\arguments{
\item{file}{Path to a single-page pdf or image file}

\item{proj_id}{Your Google Cloud Services project id.}

\item{loc}{A two-letter region code ("eu" or "us"). Defaults to "eu".}

\item{token}{An authentication token generated by \code{dai_auth()} or another auth function.}
}
\value{
a json response object
}
\description{
Sends a single-page document to Google Document AI v1beta2 for synchronous (immediate) processing. Returns a response object containing the text and additional data. This function accesses a different endpoint than the main \code{dai_sync()} function, one that has less language support, but extracts table data (which \code{dai_sync()} currently does not).
This function will be deprecated if/when the v1 endpoint incorporates table extraction.
}
\details{
Requires a Google Cloud access token (\code{google_token}) and a certain amount of configuration in RStudio; see vignettes for details. Input files can be in either .pdf, .bmp, .gif, .jpeg, .jpg, .png, or .tiff format. Extract the text from the response object with dai::text_from_dai_response(). Inspect the entire response object with httr::content().
}
\examples{
\dontrun{
response <- dai_sync("doc_page.pdf")

my_page_scan <- "001.png"
response <- dai_sync(my_page_scan)
}
}
