% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dai_token}
\alias{dai_token}
\title{Produce access token}
\usage{
dai_token(
  path = Sys.getenv("GCS_AUTH_FILE"),
  scopes = "https://www.googleapis.com/auth/cloud-platform"
)
}
\arguments{
\item{path}{path to a JSON file with a service account key}

\item{scopes}{GCS auth scopes for the token}
}
\value{
a GCS access token object (if credentials are valid) or a message (if not).
}
\description{
Produces an access token for Google Cloud Services (GCS)
}
\examples{
\dontrun{
token <- dai_token()
}
}
