% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dai_user}
\alias{dai_user}
\title{Get user info associated with service account key}
\usage{
dai_user()
}
\value{
a list of user information elements
}
\description{
Get user info associated with service account key
}
\examples{
\dontrun{
dai_user()
}
}
