% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{dai_sync_tab}
\alias{dai_async_tab}
\title{Defunct functions}
\usage{
# Deprecated in 1.0.0 -------------------------------------

dai_sync_tab(...)

dai_async_tab(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

If there's a known replacement, calling the function
will tell you about it.
}
\keyword{internal}
