% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{deprecated}
\alias{deprecated}
\title{Deprecated functions}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

If there's a known replacement, calling the function
will tell you about it.
}
\keyword{internal}
