% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Run when daiR is attached}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{name of library}

\item{pkgname}{name of package}
}
\value{
no return value, called for side effects
}
\description{
Run when daiR is attached
}
