% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Run when daiR is attached}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{name of library}

\item{pkgname}{name of package}
}
\description{
Run when daiR is attached
}
