% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{draw_blocks}
\alias{draw_blocks}
\title{Inspect block bounding boxes in Google Document AI json output}
\usage{
draw_blocks(json, dir = getwd())
}
\arguments{
\item{json}{filepath of the json output file.}

\item{dir}{path to the desired output directory.}
}
\description{
This function extracts the block bounding box coordinates
from the json file returned by Google Document AI
and plots them on an image of the original page.
}
\details{
Not vectorized, but documents can be multi-page.
Generates an annotated .png file for each page in the original
document.
}
\examples{
\dontrun{
draw_blocks("pdf_output.json")
}
}
