% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{draw_tokens}
\alias{draw_tokens}
\title{Inspect token bounding boxes in Google Document AI json output}
\usage{
draw_tokens(pdf, json)
}
\arguments{
\item{pdf}{filepath of the pdf sent for processing.}

\item{json}{filepath of the json output file.}
}
\description{
This function extracts the token bounding box coordinates
from the json file returned by Google Document AI
and plots them on a copy of the source document
for human inspection.
}
\details{
Not vectorized. Takes only pdf documents (if you processed .gif or .tiff files,
convert them to pdf first). Generates an annotated .png file for each page in the original pdf.
Assumes equal page length between the pdf and the json, so if the json output was split into shards,
split the pdf accordingly first.
}
\examples{
\dontrun{
draw_tokens("pdf_page1.pdf", "pdf_output.json")
}
}
