% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{enable_processor}
\alias{enable_processor}
\title{Enable processor}
\usage{
enable_processor(
  proc_id,
  proj_id = get_project_id(),
  loc = "eu",
  token = dai_token()
)
}
\arguments{
\item{proc_id}{a Document AI processor id.}

\item{proj_id}{a GCS project id.}

\item{loc}{a two-letter region code; "eu" or "us".}

\item{token}{an authentication token generated by
\code{dai_auth()} or another auth function.}
}
\value{
no return value, called for side effects
}
\description{
Enable processor
}
\examples{
\dontrun{
enable_processor(proc_id = get_processors()$id[1])
}
}
