% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{from_labelme}
\alias{from_labelme}
\title{Extract block coordinates from labelme files}
\usage{
from_labelme(json, page = 1)
}
\arguments{
\item{json}{a json file generated by Labelme}

\item{page}{the number of the annotated page}
}
\value{
a dataframe with location coordinates for the rectangle marked in labelme
}
\description{
Extract block coordinates from labelme files
}
\examples{
\dontrun{
new_block <- from_labelme("document1_blocks.json")
new_block <- from_labelme("document5_blocks.json", 5)
}
}
