% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_entities}
\alias{get_entities}
\title{Get entities}
\usage{
get_entities(object, type = "sync")
}
\arguments{
\item{object}{either a HTTP response object from
\code{dai_sync()} or the path to a JSON file from
\code{dai_async()}.}

\item{type}{one of "sync" or "async", depending on
the function used to process the original document.}
}
\value{
a list of dataframes, one per page
}
\description{
Extracts entities Document AI (DAI) identified by a Document AI
form parser processor.
}
\examples{
\dontrun{
entities <- get_entities(dai_sync("file.pdf"))

entities <- get_entities("file.json", type = "async")
}
}
