% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{get_ids_by_type}
\alias{get_ids_by_type}
\title{List ids of available processors of a given type}
\usage{
get_ids_by_type(
  type,
  proj_id = get_project_id(),
  loc = "eu",
  token = dai_token()
)
}
\arguments{
\item{type}{name of a processor type, e.g. "FORM_PARSER_PROCESSOR".}

\item{proj_id}{a GCS project id.}

\item{loc}{a two-letter region code; "eu" or "us".}

\item{token}{an authentication token generated by
\code{dai_auth()} or another auth function.}
}
\value{
a vector of processor ids.
}
\description{
List ids of available processors of a given type
}
\examples{
\dontrun{
get_ids_by_type("OCR_PROCESSOR")
}
}
