% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{get_processor_versions}
\alias{get_processor_versions}
\title{List available processor versions}
\usage{
get_processor_versions(
  proj_id = get_project_id(),
  proc_id = Sys.getenv("DAI_PROCESSOR_ID"),
  loc = "eu",
  token = dai_token()
)
}
\arguments{
\item{proj_id}{a GCS project id.}

\item{proc_id}{a Document AI processor id.}

\item{loc}{a two-letter region code; "eu" or "us".}

\item{token}{an authentication token generated by \code{dai_auth()} or
another auth function.}
}
\value{
a dataframe.
}
\description{
List available processor versions
}
\examples{
\dontrun{
df <- get_processor_versions()

df <- get_processor_versions(proc_id = get_processors()$id[1])
}
}
