% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R
\name{get_processors}
\alias{get_processors}
\title{List created processors}
\usage{
get_processors(proj_id = get_project_id(), loc = "eu", token = dai_token())
}
\arguments{
\item{proj_id}{a GCS project id.}

\item{loc}{a two-letter region code; "eu" or "us".}

\item{token}{an authentication token generated by \code{dai_auth()} or
another auth function.}
}
\value{
a dataframe.
}
\description{
List created processors
}
\details{
Retrieves information about the processors that
have been created in the current project and are ready for use.
For more information about processors, see the Google Document
AI documentation at
\url{https://docs.cloud.google.com/document-ai/docs}.
}
\examples{
\dontrun{
df <- get_processors()
}
}
