% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_project_id}
\alias{get_project_id}
\title{Get project id from service account file}
\usage{
get_project_id(env_var = "GCS_AUTH_FILE")
}
\arguments{
\item{env_var}{name of the .Renviron variable containing the path to the json file with your service account key.}
}
\value{
string with your project id
}
\description{
Get project id from service account file
}
\examples{
\dontrun{
project_id <- get_project_id()
}
}
