% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_tables}
\alias{get_tables}
\title{Get tables}
\usage{
get_tables(object, type = "sync")
}
\arguments{
\item{object}{either a HTTP response object from
\code{dai_sync()} or the path to a JSON file from
\code{dai_async()}.}

\item{type}{one of "sync" or "async", depending on
the function used to process the original document.}
}
\value{
a list of data frames
}
\description{
Extracts tables identified by a Document AI
form parser processor.
}
\examples{
\dontrun{

tables <- get_tables(dai_sync("file.pdf"))

tables <- get_tables("file.json", type = "async")
}
}
