% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{get_vertices}
\alias{get_vertices}
\title{Get vertices}
\usage{
get_vertices(lst)
}
\arguments{
\item{lst}{a list}
}
\description{
Helper function for extracting coordinates from DAI reponse objects
}
