% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{image_to_pdf}
\alias{image_to_pdf}
\title{Convert images to PDF}
\usage{
image_to_pdf(files, pdf_name)
}
\arguments{
\item{files}{a vector of image files}

\item{pdf_name}{name of the new pdf}
}
\description{
This function converts a vector of images to a single pdf.
}
\details{
Combines any number of image files of almost any type
to a single PDF. The vector can consist of different image file types.
See the \code{magick} package documentation for details on
supported file types. Note that on Linux, ImageMagick may not allow
conversion to pdf for security reasons. The setting can be turned off;
see: https://stackoverflow.com/a/52863413.
}
\examples{
\dontrun{
# Single file
image_to_pdf("document.jpg", "document.pdf")

# A vector of files:
image_to_pdf(images)
}
}
