% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_to_binbase}
\alias{img_to_binbase}
\title{Convert image file to base64-encoded binary tiff}
\usage{
img_to_binbase(file)
}
\arguments{
\item{file}{path to an image file}
}
\value{
string containing base64-encoded binary tiff
}
\description{
Convert image file to base64-encoded binary tiff
}
