% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_colour}
\alias{is_colour}
\title{Check that a string is a valid colour representation}
\usage{
is_colour(x)
}
\arguments{
\item{x}{a string}
}
\value{
a boolean
}
\description{
Checks whether a string is a valid colour representation.
}
\examples{
\dontrun{
is_colour("red")
is_colour("#12345")
}
}
