% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_json}
\alias{is_json}
\title{Check that a file is json}
\usage{
is_json(file)
}
\arguments{
\item{file}{a filepath}
}
\value{
boolean
}
\description{
Check that a file is json
}
\examples{
\dontrun{
is_json("file.json")
}
}
