% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pdf}
\alias{is_pdf}
\title{Check that a file is pdf}
\usage{
is_pdf(file)
}
\arguments{
\item{file}{a filepath}
}
\value{
boolean
}
\description{
Check that a file is pdf
}
\examples{
\dontrun{
is_pdf("document.pdf")
}
}
