% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf_to_binbase}
\alias{pdf_to_binbase}
\title{Convert pdf to base64-encoded binary tiff}
\usage{
pdf_to_binbase(file)
}
\arguments{
\item{file}{path to a single-page pdf file}
}
\value{
string containing base64-encoded binary tiff
}
\description{
Convert pdf to base64-encoded binary tiff
}
