% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{reassign_tokens}
\alias{reassign_tokens}
\title{Assign tokens to new blocks}
\usage{
reassign_tokens(token_df, block_df)
}
\arguments{
\item{token_df}{a dataframe generated by \code{dair::build_token_df}}

\item{block_df}{a dataframe generated by \code{dair::split_block()} or \code{dair::build_block_df}}
}
\value{
a dataframe of tokens
}
\description{
Assign tokens to new blocks
}
\examples{
\dontrun{
new_token_df <- reassign_tokens(token_df, new_block_df)
}
}
