% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{tables_from_dai_file}
\alias{tables_from_dai_file}
\title{Get tables from a Document AI json file}
\usage{
tables_from_dai_file(file)
}
\arguments{
\item{file}{filepath of a json file from DAI}
}
\value{
a list of dataframes
}
\description{
Get tables from a Document AI json file
}
\details{
Extracts all the tables that DAI identified in
the submitted document and returns them as a list of dataframes.
}
\examples{
\dontrun{
tables <- tables_from_dai_file("document.json")
}
}
