% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{tables_from_dai_response}
\alias{tables_from_dai_response}
\title{Get tables from a Document AI response object}
\usage{
tables_from_dai_response(object)
}
\arguments{
\item{object}{a json response object from DAI}
}
\value{
a list of dataframes
}
\description{
Get tables from a Document AI response object
}
\details{
Extracts all the tables that DAI identified in
the submitted document and returns them as a list of dataframes.
}
\examples{
\dontrun{
tables <- tables_from_dai_response(response)
}
}
