% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{text_from_dai_response}
\alias{text_from_dai_response}
\title{Get document text from a Document AI response object}
\usage{
text_from_dai_response(response)
}
\arguments{
\item{object}{a json response object from DAI}
}
\value{
a string.
}
\description{
This function extracts the text element of a json response object
returned from dai::dai_sync()
}
\examples{
\dontrun{
text <- text_from_dai_response(response)
}
}
