% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{transpose_block}
\alias{transpose_block}
\title{Transpose blocks}
\usage{
transpose_block(lst)
}
\arguments{
\item{lst}{a list}
}
\description{
Helper function for converting coordinates from DAI reponse objects
}
