% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{transpose_page}
\alias{transpose_page}
\title{Transpose page}
\usage{
transpose_page(x)
}
\arguments{
\item{x}{a list element in a response object converted with httr::content().}
}
\description{
Helper function for converting coordinates from DAI reponse objects
}
