% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{build_block_df}
\alias{build_block_df}
\title{Build block dataframe}
\usage{
build_block_df(json)
}
\arguments{
\item{json}{filepath of a JSON file obtained using \code{dai_async()}}
}
\value{
a block data frame
}
\description{
Creates a dataframe with the block bounding boxes
identified by Document AI (DAI) in an asynchronous request.
Rows are blocks, in the order DAI proposes to read them. Columns
are location variables such as page coordinates and page numbers.
}
\details{
The location variables are: page number, left boundary,
right boundary, top boundary, and bottom boundary.
}
\examples{
\dontrun{
block_df <- build_block_df("pdf_output.json")
}
}
