% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dai_auth}
\alias{dai_auth}
\title{Get service account token to use daiR}
\usage{
dai_auth(
  scopes = "https://www.googleapis.com/auth/cloud-platform",
  env_var = "GCS_AUTH_FILE"
)
}
\arguments{
\item{scopes}{scopes wanted for the token.}

\item{env_var}{name of the .Renviron variable containing the path to the json file with your service account key.}
}
\value{
access token
}
\description{
Get service account token to use daiR
}
\examples{
\dontrun{
# A simple call with default parameters:
dai_auth()

# A tailored call:
dai_auth(scopes = c("https://www.googleapis.com/auth/cloud-platform",
                    "https://www.googleapis.com/auth/bigquery"),
         env_var = "ANOTHER_AUTH_FILE")
}
}
