% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dai_user}
\alias{dai_user}
\title{Get user information}
\usage{
dai_user()
}
\value{
a list of user information elements
}
\description{
Fetches the Google Cloud Services (GCS) user information
associated with a service account key.
}
\examples{
\dontrun{
dai_user()
}
}
