% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{draw_blocks}
\alias{draw_blocks}
\title{Inspect block bounding boxes}
\usage{
draw_blocks(json, dir = getwd())
}
\arguments{
\item{json}{filepath of a JSON file obtained using \code{dai_async()}}

\item{dir}{path to the desired output directory}
}
\value{
no return value, called for side effects
}
\description{
Plots the block bounding boxes identified by
Document AI (DAI) onto images of the submitted document.
Generates an annotated .png file for each page in the original
document.
}
\details{
Not vectorized, but documents can be multi-page.
}
\examples{
\dontrun{
draw_blocks("pdf_output.json", dir = tempdir())
}
}
