% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_project_id}
\alias{get_project_id}
\title{Get project id from service account file}
\usage{
get_project_id(path = Sys.getenv("GCS_AUTH_FILE"))
}
\arguments{
\item{path}{path to the json file with your service account key.}
}
\value{
a string with a project id
}
\description{
Get project id from service account file
}
\examples{
\dontrun{
project_id <- get_project_id()
}
}
