% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_project_id}
\alias{get_project_id}
\title{Get project id}
\usage{
get_project_id(path = Sys.getenv("GCS_AUTH_FILE"))
}
\arguments{
\item{path}{path to the JSON file with your service account key}
}
\value{
a string with a GCS project id
}
\description{
Fetches the Google Cloud Services (GCS) project id
associated with a service account key.
}
\examples{
\dontrun{
project_id <- get_project_id()
}
}
