% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{img_to_binbase}
\alias{img_to_binbase}
\title{Image to base64 tiff}
\usage{
img_to_binbase(file)
}
\arguments{
\item{file}{path to an image file}
}
\value{
a base64-encoded string
}
\description{
Converts an image file to a base64-encoded binary .tiff file.
}
\examples{
\dontrun{
img_encoded <- img_to_binbase("image.png")
}
}
