% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_json}
\alias{is_json}
\title{Check that a file is JSON}
\usage{
is_json(file)
}
\arguments{
\item{file}{a filepath}
}
\value{
a boolean
}
\description{
Checks whether a file is a JSON file.
}
\examples{
\dontrun{
is_json("file.json")
}
}
