% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_pdf}
\alias{is_pdf}
\title{Check that a file is PDF}
\usage{
is_pdf(file)
}
\arguments{
\item{file}{a filepath}
}
\value{
a boolean
}
\description{
Checks whether a file is a PDF file.
}
\examples{
\dontrun{
is_pdf("document.pdf")
}
}
