% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf_to_binbase}
\alias{pdf_to_binbase}
\title{PDF to base64 tiff}
\usage{
pdf_to_binbase(file)
}
\arguments{
\item{file}{path to a single-page pdf file}
}
\value{
a base64-encoded string
}
\description{
Converts a PDF file to a base64-encoded binary .tiff file.
}
\examples{
\dontrun{
doc_encoded <- pdf_to_binbase("document.pdf")
}
}
