% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output.R
\name{reassign_tokens}
\alias{reassign_tokens}
\title{Assign tokens to new blocks}
\usage{
reassign_tokens(token_df, block_df)
}
\arguments{
\item{token_df}{a dataframe generated by \code{build_token_df()}}

\item{block_df}{a dataframe generated by \code{dair::split_block()}
or \code{dair::build_block_df()}}
}
\value{
a token data frame
}
\description{
This is a specialized function for use in connection
with text reordering. It modifies a token dataframe by assigning
new block bounding box values to a subset of tokens based on
prior modifications made to a block dataframe.
}
\details{
The token and block data frames provided as input must be
from the same JSON output file.
}
\examples{
\dontrun{
new_token_df <- reassign_tokens(token_df, new_block_df)
}
}
