% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{reassign_tokens2}
\alias{reassign_tokens2}
\title{Assign tokens to a single new block}
\usage{
reassign_tokens2(token_df, block, page = 1)
}
\arguments{
\item{token_df}{a data frame generated by \code{dair::build_token_df}}

\item{block}{a one-row data frame of the same format as \code{token_df}}

\item{page}{the number of the page on which the block belongs}
}
\value{
a token data frame
}
\description{
This is a specialized function for use in connection
with text reordering. It is designed to facilitate manual splitting
of block boundary boxes and typically takes a one-row block dataframe
generated by \code{from_labelme()}.
}
\examples{
\dontrun{
new_token_df <- reassign_tokens2(token_df, new_block_df)
new_token_df <- reassign_tokens2(token_df, new_block_df, 5)
}
}
