% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{tables_from_dai_file}
\alias{tables_from_dai_file}
\title{Get tables from output file}
\usage{
tables_from_dai_file(file)
}
\arguments{
\item{file}{filepath of a JSON file obtained using \code{dai_async_tab()}}
}
\value{
a list of data frames
}
\description{
Extracts all the tables that Document AI (DAI)
identified in an asynchronous processing request.
}
\examples{
\dontrun{
tables <- tables_from_dai_file("document.json")
}
}
