% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{tables_from_dai_file}
\alias{tables_from_dai_file}
\title{Get tables from output file}
\usage{
tables_from_dai_file(file)
}
\arguments{
\item{file}{filepath of a JSON file obtained using \code{dai_async_tab()}}
}
\value{
a list of data frames
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{tables_from_dai_file()} is deprecated; please use \code{get_text()} instead.
}
\examples{
\dontrun{
tables <- tables_from_dai_file("document.json")
}
}
