% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{tables_from_dai_response}
\alias{tables_from_dai_response}
\title{Get tables from response object}
\usage{
tables_from_dai_response(object)
}
\arguments{
\item{object}{an HTTP response object returned by \code{dai_sync_tab()}}
}
\value{
a list of data frames
}
\description{
Extracts all the tables that Document AI (DAI)
identified in a synchronous processing request.
}
\examples{
\dontrun{
tables <- tables_from_dai_response(response)
}
}
