% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{tables_from_dai_response}
\alias{tables_from_dai_response}
\title{Get tables from response object}
\usage{
tables_from_dai_response(object)
}
\arguments{
\item{object}{an HTTP response object returned by \code{dai_sync_tab()}}
}
\value{
a list of data frames
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{tables_from_dai_response()} is deprecated; please use \code{get_tables()} instead.
}
\examples{
\dontrun{
tables <- tables_from_dai_response(response)
}
}
