% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{text_from_dai_file}
\alias{text_from_dai_file}
\title{Get document text from a Document AI json file}
\usage{
text_from_dai_file(file)
}
\arguments{
\item{file}{filepath of a json file from DAI.}
}
\value{
a string.
}
\description{
This function extracts the text element of a json output file
generated by Google Document AI.
}
\examples{
\dontrun{
text <- text_from_dai_file("output.json")
}
}
