% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{text_from_dai_file}
\alias{text_from_dai_file}
\title{Get text from output file}
\usage{
text_from_dai_file(file)
}
\arguments{
\item{file}{filepath of a JSON file obtained using \code{dai_async()}}
}
\value{
a string
}
\description{
Extracts the text OCRed by Document AI (DAI) in an
asynchronous processing request.
}
\examples{
\dontrun{
text <- text_from_dai_file("output.json")
}
}
