% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_output.R
\name{text_from_dai_response}
\alias{text_from_dai_response}
\title{Get text from HTTP response object}
\usage{
text_from_dai_response(response)
}
\arguments{
\item{response}{an HTTP response object returned by \code{dai_sync()}}
}
\value{
a string
}
\description{
Extracts the text OCRed by Document AI (DAI) in a
synchronous processing request.
}
\examples{
\dontrun{
text <- text_from_dai_response(response)
}
}
